infrastructure/lambda/api-handler/requirements.txt
boto3==1.28.0
pyyaml==6.0.1
